Add support for multiple bound MySQL services
There may be more than one MySQL service bound to an application. This change updates confic.inc.php such that an array of MySQL servers is populated using the services bound as extracted from the VCAP_SERVICES environment variable. Also, all services named either 'cleardb' or starting with the string 'mysql' are used. The latter is because implementations such as Bluemix bind local MySQL services with names such as 'mysql-5.5'.